Extract ImapStoreSettings from ImapStore
