+++ k9mail/src/main/java/com/fsck/k9/activity/setup/AccountSetupIncoming.java
@@ -28,7 +28,7 @@ import com.fsck.k9.mail.ServerSettings.Type;
 import com.fsck.k9.mail.Store;
 import com.fsck.k9.mail.Transport;
 import com.fsck.k9.mail.store.RemoteStore;
-import com.fsck.k9.mail.store.imap.ImapStore.ImapStoreSettings;
+import com.fsck.k9.mail.store.imap.ImapStoreSettings;
 import com.fsck.k9.mail.store.webdav.WebDavStore.WebDavStoreSettings;
 import com.fsck.k9.account.AccountCreator;
 import com.fsck.k9.service.MailService;
